refactor(#741): refactor searchController
