--- api/controllers/v1/SearchController.js
@@ -1,14 +1,0 @@
-/**
- */
-
-const searchController = require('../SearchController');
-
-module.exports = {
-  search: (req, res) => {
-    searchController.search(req, res);
-  },
-
-  advancedSearch: (req, res) => {
-    searchController.advancedSearch(req, res);
-  },
-};
